created db, routes, corrected file name gigs.js to Gig.js, moved code from gigs.js to routes/gigs.js,created public folder
