+++ .env.EXAMPLE
@@ -1,0 +1,3 @@
+DB_NAME= 'gig_db'
+DB_PASSWORD= ''
+DB_USER= ''
