+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "codegig",
+  "version": "1.0.0",
+  "description": "Job finder APP for Coders",
+  "main": "app.js",
+  "scripts": {
+    "test": "none",
+    "start": "node app.js",
+    "dev": "nodemon app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Corey96/python-gig.git"
+  },
+  "author": "Andrea, Corey, Nadira, Saniya",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Corey96/python-gig/issues"
+  },
+  "homepage": "https://github.com/Corey96/python-gig#readme",
+  "dependencies": {
+    "bcrypt": "^5.0.1",
+    "body-parser": "^1.20.0",
+    "bootstrap": "^5.1.3",
+    "express": "^4.17.3",
+    "express-handlebars": "^6.0.3",
+    "mysql2": "^2.3.3",
+    "nodemon": "^2.0.15",
+    "pg": "^8.7.3",
+    "sequelize": "^6.18.0"
+  }
+}
