+++ shema.sql
@@ -1,0 +1,4 @@
+DROP DATABASE IF EXISTS gig_db;
+
+-- CREATE DATABASE
+CREATE DATABASE gig_db;
