+++ views/layouts/main.handlebars
@@ -9,7 +9,7 @@
     crossorigin="anonymous">
 
 
-  <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="/css/style.css">
   <title>CodeGig</title>
 </head>
